+++ lib/extensions/persistent_session/pow_persistent_session.ex
@@ -1,0 +1,7 @@
+defmodule PowPersistentSession do
+  @moduledoc false
+  use Pow.Extension.Base
+
+  @impl true
+  def phoenix_controller_callbacks?(), do: true
+end