+++ test/mix/tasks/extension/ecto/pow.extension.ecto.gen.migrations_test.exs
@@ -1,4 +1,12 @@
 defmodule Mix.Tasks.Pow.Extension.Ecto.Gen.MigrationsTest do
+  # Implementation needed for `Pow.Extension.Base.has?/2` check
+  defmodule ExtensionMock do
+    use Pow.Extension.Base
+
+    @impl true
+    def ecto_schema?(), do: true
+  end
+
   defmodule ExtensionMock.Ecto.Schema do
     use Pow.Extension.Ecto.Schema.Base