+++ CHANGELOG.md
@@ -10,6 +10,12 @@
   * `PowPersistentSession.Plug.Cookie.create/3` will now, instead of adding `:session_fingerprint` to the metadata, populate the `:session_metadata` keyword list with `:fingerprint`
   * `PowPersistentSession.Plug.Cookie.authenticate/2` will now populate session metadata with what exists in `:session_metadata` key for the persistent session metadata
 * `PowPersistentSession.Plug.Cookie.create/3` now ensures to delete the previous persistent session first, if one is found in cookies
+* Extensions are now expected to have a base module with compile-time information whether certain modules are available to prevent unnecessary `Code.ensure_compiled?/1` calls:
+  * Added `Pow.Extension.Base` module
+  * Added `PowEmailConfirmation` module
+  * Added `PowInvitation` module
+  * Added `PowPersistentSession` module
+  * Added `PowResetPassword` module
 
 ## v1.0.14 (2019-10-29)